Refactor: Improve Theming, JS, and Component Structure
This commit addresses several inconsistencies and improves the overall structure of the application.

- **JavaScript Refactoring**:
  - Modularized the monolithic `script.js` into `theme.js`, `ui.js`, and `constants.js` for better organization and maintenance.
  - Created a central `main.js` to handle client-side script initialization.
  - Ensured all DOM-accessing code is client-side only to prevent SSR errors, resolving the `document is not defined` issue.

- **Styling and Theming**:
  - Unified the color scheme across the application by replacing hardcoded colors with Tailwind CSS theme variables.
  - Corrected conflicting background styles in the main layout.
  - Made the chat message bubbles in `messaging.astro` theme-aware.

- **Componentization**:
  - Replaced the duplicated homepage cards with a single, reusable `ModeCard.astro` component.

- **Asset and Link Handling**:
  - Replaced external and placeholder images with local assets in `public/images/`.
  - Added descriptive `alt` text to all images.
  - Added a `404.astro` page for better error handling.
  - Removed a non-functional "Search" link from the sidebar.

- **CI/CD**:
  - Reviewed and confirmed that the GitHub Actions workflow is up-to-date and has the correct permissions.

- **Testing**:
  - All existing tests pass.
  - Added a Playwright script to visually verify the theme-switching functionality and the new component structure.
