+++ src/components/DesktopSidebar.astro
@@ -1,18 +1,19 @@
 ---
 ---
-<aside class="hidden lg:flex lg:flex-col lg:w-64 bg-white dark:bg-gray-800 shadow-lg">
-    <div class="flex items-center justify-center h-20 shadow-md">
+<aside class="hidden lg:flex lg:flex-col lg:w-64 bg-community-bg dark:bg-hookups-bg shadow-lg border-r border-white/5 dark:border-white/5">
+    <div class="flex items-center justify-center h-20 shadow-md bg-white/5 dark:bg-white/5">
         <h1 class="text-2xl font-bold text-community-accent dark:text-hookups-accent">Limpopo C.</h1>
     </div>
     <nav class="flex-1 px-4 py-8 space-y-2">
-        <a href="/" class="flex items-center px-4 py-2 text-gray-700 dark:text-gray-200 hover:bg-gray-200 dark:hover:bg-gray-700 rounded-md">Home</a>
-        <a href="#" class="flex items-center px-4 py-2 text-gray-700 dark:text-gray-200 hover:bg-gray-200 dark:hover:bg-gray-700 rounded-md">Search</a>
-        <a href="/community/" class="flex items-center px-4 py-2 text-gray-700 dark:text-gray-200 hover:bg-gray-200 dark:hover:bg-gray-700 rounded-md">Events</a>
-        <a href="/locations/" class="flex items-center px-4 py-2 text-gray-700 dark:text-gray-200 hover:bg-gray-200 dark:hover:bg-gray-700 rounded-md">Locations</a>
-        <a href="/messaging/" class="flex items-center px-4 py-2 text-gray-700 dark:text-gray-200 hover:bg-gray-200 dark:hover:bg-gray-700 rounded-md">Messages</a>
-        <a href="/profile/" class="flex items-center px-4 py-2 text-gray-700 dark:text-gray-200 hover:bg-gray-200 dark:hover:bg-gray-700 rounded-md">Profile</a>
+        <a href="/" class="flex items-center px-4 py-2 rounded-md text-community-text dark:text-hookups-text hover:bg-community-accent/10 dark:hover:bg-hookups-accent/10">Home</a>
+        <a href="/community/" class="flex items-center px-4 py-2 rounded-md text-community-text dark:text-hookups-text hover:bg-community-accent/10 dark:hover:bg-hookups-accent/10">Events</a>
+        <a href="/locations/" class="flex items-center px-4 py-2 rounded-md text-community-text dark:text-hookups-text hover:bg-community-accent/10 dark:hover:bg-hookups-accent/10">Locations</a>
+        <a href="/messaging/" class="flex items-center px-4 py-2 rounded-md text-community-text dark:text-hookups-text hover:bg-community-accent/10 dark:hover:bg-hookups-accent/10">Messages</a>
+        <a href="/profile/" class="flex items-center px-4 py-2 rounded-md text-community-text dark:text-hookups-text hover:bg-community-accent/10 dark:hover:bg-hookups-accent/10">Profile</a>
     </nav>
     <div class="mt-auto p-4">
-        <button id="theme-toggle-btn" class="w-full py-2 px-4 bg-gray-200 dark:bg-gray-700 rounded-lg text-gray-800 dark:text-white">Toggle Theme</button>
+        <button id="theme-toggle-btn" class="w-full py-2 px-4 rounded-lg bg-community-accent/20 dark:bg-hookups-accent/20 text-community-text dark:text-hookups-text hover:bg-community-accent/30 dark:hover:bg-hookups-accent/30">
+            Toggle Theme
+        </button>
     </div>
 </aside>
