+++ src/components/ModeCard.astro
@@ -1,0 +1,32 @@
+---
+interface Props {
+  title: string;
+  description: string;
+  imageUrl: string;
+  imageAlt: string;
+  link: string;
+  accentColor: 'community' | 'hookups';
+}
+
+const { title, description, imageUrl, imageAlt, link, accentColor } = Astro.props;
+
+const accentClasses = {
+  community: 'text-community-accent bg-community-accent',
+  hookups: 'text-hookups-accent bg-hookups-accent',
+};
+
+const selectedAccent = accentClasses[accentColor];
+---
+
+<div class="group w-full max-w-sm rounded-2xl shadow-lg transition-all duration-300 hover:scale-105 hover:shadow-2xl overflow-hidden bg-white/10 backdrop-blur-lg border border-white/20">
+  <img src={imageUrl} alt={imageAlt} class="w-full h-48 object-cover" />
+  <div class="p-6 flex flex-col justify-between flex-1">
+    <div>
+      <h2 class:list={["text-3xl font-bold mb-2", selectedAccent.split(' ')[0]]}>{title}</h2>
+      <p class="text-gray-200 mb-4">{description}</p>
+    </div>
+    <a href={link} class:list={["mt-auto px-8 py-3 text-white font-bold rounded-full shadow-lg transform transition-transform duration-300 group-hover:scale-110 self-center", selectedAccent.split(' ')[1]]}>
+      {accentColor === 'hookups' ? 'Enter' : 'Explore'}
+    </a>
+  </div>
+</div>
