+++ src/pages/404.astro
@@ -1,0 +1,16 @@
+---
+import MainLayout from '../layouts/MainLayout.astro';
+---
+
+<MainLayout title="404: Not Found">
+  <div class="flex flex-col items-center justify-center text-center h-full p-8">
+    <h1 class="text-8xl font-bold text-community-accent dark:text-hookups-accent">404</h1>
+    <p class="text-2xl font-semibold mt-4">Page Not Found</p>
+    <p class="mt-2 text-lg text-community-text dark:text-hookups-text">
+      Sorry, we couldn't find the page you were looking for.
+    </p>
+    <a href="/" class="mt-8 px-6 py-3 bg-community-accent dark:bg-hookups-accent text-white font-bold rounded-full shadow-lg hover:scale-105 transition-transform">
+      Go Back Home
+    </a>
+  </div>
+</MainLayout>
