+++ src/pages/index.astro
@@ -1,41 +1,31 @@
 ---
 import MainLayout from '../layouts/MainLayout.astro';
+import ModeCard from '../components/ModeCard.astro';
 ---
 <MainLayout title="Limpopo Connect">
-    <div class="relative min-h-full flex items-center justify-center bg-gradient-to-br from-purple-600 via-pink-500 to-red-500 animate-gradient-xy">
+    <div class="relative min-h-full flex items-center justify-center p-4 sm:p-8 bg-gradient-to-br from-purple-600 via-pink-500 to-red-500">
         <div class="absolute bg-black opacity-50 inset-0"></div>
-        <div class="relative z-10 text-center text-white p-8">
+        <div class="relative z-10 text-center text-white w-full">
             <h1 class="text-5xl md:text-7xl font-bold mb-4 tracking-tight">Connect Your Way</h1>
             <p class="text-xl md:text-2xl mb-10 font-light">One platform, two worlds. Find your connection.</p>
 
-            <div class="flex flex-col md:flex-row items-center justify-center gap-6">
-                <!-- Hookups Mode Card -->
-                <div class="group relative w-80 h-96 rounded-2xl shadow-lg transition-all duration-300 hover:scale-105 hover:shadow-2xl overflow-hidden" style="background-image: url('https://i.imgur.com/TyhAOt9.png'); background-size: cover; background-position: center;">
-                    <div class="absolute inset-0 bg-black bg-opacity-50"></div> <!-- Overlay -->
-                    <div class="relative z-10 p-6 flex flex-col justify-between h-full text-white">
-                        <div>
-                            <h2 class="text-3xl font-bold text-hookups-accent mb-2">Hookups</h2>
-                            <p class="text-gray-200 mb-4">Discreet. Anonymous. Direct.</p>
-                        </div>
-                        <a href="/hookups/" class="mt-auto px-8 py-3 bg-hookups-accent text-white font-bold rounded-full shadow-lg transform transition-transform duration-300 group-hover:scale-110 self-center">
-                            Enter
-                        </a>
-                    </div>
-                </div>
-
-                <!-- Community Mode Card -->
-                <div class="group w-80 h-96 bg-white bg-opacity-20 rounded-2xl shadow-lg backdrop-blur-lg border border-white border-opacity-20 transition-all duration-300 hover:scale-105 hover:shadow-2xl overflow-hidden">
-                    <img src="https://via.placeholder.com/320x200/FF9500/F5F5F7?text=Vibrant" alt="Community" class="w-full h-1/2 object-cover">
-                    <div class="p-6 flex flex-col justify-between flex-1">
-                        <div>
-                            <h2 class="text-3xl font-bold text-community-accent mb-2">Community</h2>
-                            <p class="text-gray-200 mb-4">Events. Groups. Culture.</p>
-                        </div>
-                        <a href="/community/" class="mt-auto px-8 py-3 bg-community-accent text-white font-bold rounded-full shadow-lg transform transition-transform duration-300 group-hover:scale-110 self-center">
-                            Explore
-                        </a>
-                    </div>
-                </div>
+            <div class="grid grid-cols-1 md:grid-cols-2 gap-8 max-w-4xl mx-auto">
+                <ModeCard
+                    title="Hookups"
+                    description="Discreet. Anonymous. Direct."
+                    imageUrl="/images/hookups-card.png"
+                    imageAlt="A stylized, dark image representing privacy and connection"
+                    link="/hookups/"
+                    accentColor="hookups"
+                />
+                <ModeCard
+                    title="Community"
+                    description="Events. Groups. Culture."
+                    imageUrl="/images/community-card.png"
+                    imageAlt="A vibrant image representing community events and gatherings"
+                    link="/community/"
+                    accentColor="community"
+                />
             </div>
         </div>
     </div>
