+++ src/scripts/constants.js
@@ -1,0 +1,10 @@
+/**
+ * @file Centralized constants for the application.
+ * @description Defines keys for localStorage and other site-wide constants
+ * to avoid magic strings and ensure consistency.
+ */
+
+export const STORAGE_KEYS = {
+  USER_PROFILE: 'limpopoConnectUserProfile',
+  THEME: 'theme',
+};
